+++ app/models/sla_status_query.rb
@@ -76,7 +76,7 @@ class SlaStatusQuery < Query
 
   # For Query Class
   def base_scope
-    self.queried_class.visible.where(statement)
+    self.queried_class.visible.with_references.where(statement)
   end
 
   def all_sla_types_values